Make json_stream_sampler handle UID, GID, and perm
Without the patch, the json_stream_sampler ignores the UID, GID, and
perm given at the config line.

With the patch, the plugin creates LDMS sets with the specified UID,
GID, and permission. If one of the values are not given, it will use
LDMS's default UID, GID, and Perm values as the other sets.

The S_uid, S_gid, S_perm are meta metrics that store the stream data's
UID, GID, and permission, respectively. The values of the meta metrics
change according to the received stream data. However, the set's
security (UID, GID, and perm) do not change after the set creation.
Admins can change the set security using the config command set_sec_mod
as they will do when modifying the security of any LDMS sets.