Add bridge and fix passive connections
The patch fixes the passive code path in LDMS and LDMSD, and add the
bridge_add, bridge_del, bridge_start, bridge_stop commands in
ldmsd_controller and configuration files. The commands are a wrapper of
the corresponding producer config commands.

Under the hood, LDMSD treats bridges as another producer type. It does
not have the bridge config objects.