Cache the Linux Thread ID of Zap threads
The patch makes Zap cache the Linux Thread ID besides the POSIX Thread
ID to include the ID with the thread statistics. The Linux Thread ID can
be used to explore the thread information in /proc and in conjuction
with the top command's results.

The patch also adds a new API, zap_ep_thread_id(), to return the Linux
Thread ID.