Modify the releasing Zap endpoints from IO thread process
The patch separates the process of reducing the number of Zap endpoints
associated with an IO thread from the process of releasing the endpoint
from the IO thread. Additionally, it defers the nullification of
`zap->thread` until the endpoint reference reaches zero.

`zap_sock` and `zap_ugni` need to remove their events from epoll before
delivering the disconnected event to applications, preventing the
reception of additional epoll events. However, without this patch, this
results in resetting `zap->thread` to NULL. Thus, the applications
cannot retrive Zap thread information in the disconnected path. The
patch addresses this limitation.