rdc_sampler: Add configure check for librdc_ras
ROCm likes to put librdc_ras in a different library path than librdc_bootstrap,
but librdc_bootstrap assumes that librdc_ras is in the library search path
for dlopen().

Add a configure check for librdc_ras, and use the resulting
build variables in rdc_sampler. This creates a configure option
--with-librdc_ras-prefix so librdc_ras path can be specified
separately.