Invalidate RDMA memory descriptors on set delete
Memory descriptors were previously invalidated when the last
set reference was removed. This resulted in previously scheduled
updates succeeding on a set that had been deleted. If this set memory
had been reused in the interum, garbage would be read into a set
resulting in a crash in the update complete callback.