Fix `char *` conversion in ldms.pyx
`BYTES()` converted `str`, or `object` into `bytes` which can later
converted to `char *` by Cython. However, `BYTES(None)` got converted to
`b'None'`, which is unintended. What we wanted in `ldms.pyx` is to
convert `None` as NULL. In order to do so, we need `CSTR()` to convert
`bytes` or `None` into `char *` or `NULL` respectively. We cannot
convert `object` into `char *` or `NULL` in one go in `CSTR` because the
first `bytes` conversion would be an object that temporarily lives in
the `CSTR()` function and the `<char*>` conversion on that temporary
object will be an invalid pointer.

This patch also add `CBYTES(c_str)` that returns `None` when `c_str` is
`NULL`. We need this because `char *` is automatically converted to
`bytes` automatically by Cython when it is assigned to a Python
variable, but the program crashes when the pointer is NULL (due to the
subsequent invalid `PyBytes_FromString(NULL)` call).