Refactor & improve zap to use ovis_thrstats
Without the patch, Zap collects the utilization statistics per number of
samples not a fixed time window length. The result may not agreed with
`top` output as `top` reports per a fixed time window.

The patch changes this so that the reported Zap thread usage is the data
of the last 3 seconds.