Add sampler plugin 'ldmscon2024_sampler'
The sampler plugin is an example and a reference point from the
Sampler Plugin Implement focus group at LDMSCON2024.

The commit infomation also shows the necessary files to be modified and
created in order to adding a new sampler plugin to the git tree.

The sampler plugin is disabled by default as it is not for production
environment. To build the plugin, --enable-ldmscon2024_sampler must be
given when ./configure.