Fix 4.3.3-compat-test failed in IPv6 environment
If we use `AF_INET`, `localhost` is resolved to '127.0.0.1', but
in IPv6 environment, 'localhost' is resolved to '::1'. Since `ldmsd` 4.3
only supported IPv4, the `4.3.3-compat-test` failed as the top-of-tree
`ldmsd` (and `ldms_ls`) tried to connect to '::1' that was not listened
by `ldmsd` 4.3. This patch explcitly gives '127.0.0.1' host address to
address the issue.