Also add '::ffff:127.0.0.1' loopback comparison for IPv6
'::1' is the loopback for IPv6 family. The variable `in6addr_loopback`
defined in "netinet/in.h" was set to `::1` value for easy comparison.
However, we missed `::ffff:127.0.0.1` which is the IPv4 loopback in the
IPv6 format. This patch uses both `::1` and `::ffff:127.0.0.1` to
determine if an IPv6 address is a loopback.