Workaround for GitHub workflow dropping CentOS7 support
CentOS7 is old, and GitHub Runner recently uses 'node20' which requires
newer GLIBC that does not exist in CentOS7. As a result, the three
workflows in our repository that use 'ovis-centos-build' image fails to
even checkout the code since `actions/checkout@v3` was implemented in
TypeScript that run using 'node.js'.

According to this:
https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/
CentOS7 is also removed from the supported OS list due to node20 not
supported on CentOS7.

This patch does the following on the CentOS7-based workflows:

- Use `actions/checkout@v1` (instead of v3) because it is implemented as
  a plugin that comes with the GitHub Runner (C++). We do not use any
  advanced features that does not exist in v1.

- Remove the use of `actions/upload-artifact`. v3 is implemented in
  TypeScript, and v1 (C++ Runner Plugin) does not work. The artifact's
  purpose was for debugging.