decomp_static: correct timestamp calculations
The timestamp diff and mean operations incorrectly
treated the sec and usec fields separately. The fields
need to be considered together when applying math.

Make diff_op and diff_mean operations correct by combining the
parts of the timesamp (sec and usec) into a single number before
applying the calulation.