Perform validity checks on set instance name
The set name provided to ldms_set_create cannot contain characters
that cannot be encoded in a quoted string. Prior to this patch it
was possible to create a set that would crash ldms_xprt_dir because
the result is encoded as a JSON object and the instance name as a
JSON string.