static decomposition: yield a config error for "op" without "group"
When "op" was present in the static decomposition configuration but
"group" was not defined, the static decomposition skipped row cache
creation and continue working with `row_cache` being `NULL`.
Consequently, when `decompose()` was later called, `ldmsd` terminated
with SIGSEGV.

This patch add a sanity check in `config()` so that the static
decomposition yields an error when "group" is missing in the presence of
"op".