Support for cli options in YAML file
Keys in the daemons section aside from "endpoints", "names", "hosts", and "environment" are assumed to be CLI commands or configuration commands. Keys with a string or integer value will be treated as CLI commands, while keys that have a dictionary as their value are treated as configuration commands.
Update error handling when a configuration command/argument is invalid