Validate set schema for subsequent calls to SOS store
It is possible that a set could be provided to the SOS store()
function that has a different schema than what was originally
provided. This can crash the ldmsd and/or store incorrect
data to the container.

This change caches the set digest of the first set in the
storage instance and compares this cached value to the set
digest provided in the store() function.

If these digests do not match, the set is ignored and
not stored.