Fix ovis_event thrstat calculation
When `__thrstat_wait_start()` was called, the caller was going into
INACTIVE state from ACTIVE state. Hence, the duration from `wait_end` to
`now` was the active time and should be added to `proc_tot`.

Likewise, when `__thrstat_wait_end()` was called, the caller just woke
up. Hence, the duration from `wait_start` to `now` was an INACTIVE time
and should be added to `wait_tot`.