Add re-entrant context support for `decompose(strgp, set, ...)`
Add a support for decomposition to set a re-entrant context for each
strgp-set pair in `decompose()` interface. In other words, the next
`decompose()` call of the same `strgp` and `set` will have the same
`context` set in the current call of `decompose()` of the same `strgp`
and `set`. This enables decomposition plugins to efficiently manage
decomposition resources associated with strgp-set without looking them
up in every `decompose()` call.