Bug fix for updater parsing
Addresses bug when "updaters" is omitted from "peers" dictionary in top level "aggregators" dictionary.
Fixes bug that occurs when "producers" regular expression is not defined in the "updaters" dictionary.
Updaters now correctly only add the parent dictionary's producers when "produers" is not defined in the updaters dictionary.
Updated producer balancing when multiple aggregators share producers in the YAML configuration.