Fix store_csv header for timestamp data type
In store_csv, a timestamp type metric always produces 2 output columns:
floating-point seconds (or floating-point milli seconds depending on the
configuration) and integer microseconds. The "Timestamp" metric was
handled correctly by having 2 header columnns produced for the
"Timestamp" and "Timestamp_usec". However, for other metrics of type
timestamp, only one header column was produced for the metric, making
a mismatch between the headers and the data. This patch applies the
logic for "Timestamp" metric to other metrics of timestamp data type.