The config <-> term handling in many plugins is broken
Many plugins initialize global variables in their get_plugin()
and/or config() routines. Many of these global variables are
free() in the term() handler, however, they are not reset to
NULL. A subsequent call to get_plugin() and/or config() will
end up using a previously freed global variable.

This patch addresses the issue with ovis_log_destroy(), however,
many of the plugins have other global variables that need to be
addressed.