Fix SIGSEGV in case of invalid log level (#1660)
In the case of single log level, the function find the matching level
from the `level_bl`, but it forgot to check the breeaking `NULL` entry
and causing segmentation fault. In addition, the return code of the
single-level case is also changed to `-EINVAL` to match that of the
multi-level case (which matched the test cases).