Remove "name" from the plugin API structs
ldmsd uses the name of plugin library to load a plugin, not the
name in the plugin structure. Previously, it would then use
the name from the api struct in error messages, not the plugin
name from the "load" command and the library name. This created
an oppurtunity for mismatched name reporting, and relied on
human diligence to keep them in sync.

Since we do not use the name from the API struct at plugin
load time, and since a plugin can now get its plugin name through
ldmsd_plug_name_get(), we no longer need the name in the API
struct.

Thereforce, we remove the name from the API struct.

Closes #1691