Change method to handle term in sampler and store
This change modifies how the term command handling
determines if plugin is still in use.

For samplers the 'os' attribute is checked. This attribute is
set when the sampler is started and reset to zero when the sampler
is stopped.

For stores, a new list is maintained that maintains all the
storage policies that refer to this store. When a storage
policy is added, it is added to this list. When the storage
policy is deleted, the policy is removed from the list.