Allow empty string value for strgp container=
We intentionally do not check if the value is an emptry string.
With stores, like store_avro_kafka, the server configuration is more
complicated than a single field, and happens in the store, not in
the strgp container= line. This function needs to allow returning
an emptry string value to support that behavior.