add 'plugin_keywords' to yaml config dictionary
This lets the user specify
```
config:
  foo: bar
  plugin_keywords: ["timing", "bob"]
or
config:
  foo: bar
  plugin_keywords: timing
```
to support the existing ldmsd configuration language which allows keywords
(as distinct from attr=value pairs).
The results of the above would be
config ... foo=bar timing bob
or
config ... foo=bar timing

The name 'plugin_keywords' is reserved for ldmsd yaml use, as it seems unlikely
to be used by future plugins. It seems, on the other hand, that some future
plugin writer could easily come up with a parameter 'keywords'.