Remove non-existent plugin types
Remove from "enum ldmsd_plugin_type" (inside struct ldmsd_plugin)
the plugin types that do no currently exist. None of the following
are valid plugin types, of type "struct ldmsd_plugin" at this time:

     LDMSD_PLUGIN_AUTH,
     LDMSD_PLUGIN_DECOMP

The only valid plugin types at this time are:

    LDMSD_PLUGIN_SAMPLER
    LDMSD_PLUGIN_STORE

But we also leave in:

     LDMSD_PLUGIN_OTHER

just because b4.4 had it, and our CI/CD checks are complaining when it
is remove.