Use sampler name, rather than system hostname when generating v4 configuration
Changes the parser to use the sampler name as the producer and instance names when generating
v4 configuration files with ldmsd_yaml_parser.
Before this patch if a user attempted to generate configuration files for multiple samplers on a single machine
and failed to specify producer/instance names, the producer would be set to the hostname of the machine that
generated the configuration file.