cxi_sampler: Introduce the virtual metric derived:link_restarts
Introduce the virtual metric derived:link_restarts.  When specified
in tel_counters or tel_counters_file, the cxi_sampler watches
the link_restarts timestamp files and keeps a count of the changes
in the metric named "derived:link_restarts". This allows us to
count up to 10 link restarts per sample interval, which is good
enough in practice for relatively quick sampling rates (say
5s or faster) to detect floods of link restart issues.