Fix the default value of rail, quota, rx_rate of advertised producers
Without the patch, the rail, quota, and rx_rate values in the
prdcr_status output of advertised producers are 0 if a value isn't
specified at prdcr_listen line, which is inaccurate.

The patch is to make prdcr_status report the correct values, and it does
not affect how ldmsd setup the connection between the peer daemon and
the aggregator.