Fix ovis_thrstats_result_get()
If the given interval is 0, the function always determines that there is
not enough data points. The function will also determine that there is not
enough data poits if the given interval is larger than the duration the
data points have been collected.

The patch changes this behavior so that the function will return the
thread statistics of the maximum duration it has the data points.