lustre: Share updated stats reading function
Backport of 83893e9b5ed1e7cd84d8f933ededc93820a186b8

We backport the new implementation of the stats reading function
into each of the three lustre samplers, but we don't share the
function like on main. The sharing involved more moving around of
code than we want to make on b4.4.