`dispatch-container-bake` workflow on release event
Move `dispatch-container-bake` job into its own workflow and change
the event the workflow depends on from `on push tags` to `on release
published` so that when a release is created (by any means; not just by
matching pushed tag) our workflow will dispatch the container bake
workflow in `narategithub/ldms-containers` repository.

Also conditioned this workflow to run only on `ovis-hpc/ldms` or
`narategithub/ldms`, and change SOS branch to `b2.6`.

[fixup] dispatch