lustre: Remove SAMP define, add initial context structs
Remove the SAMP define from each of the plugins.

In the process, we want to know the plugin instance (config) name,
so we introduce a per-instance context struct to pass the data
through the necessary call stacks.

Most of the globals are not yet moved to the new context structs,
so the plugins do not yet have multi-instance support.