Fix socket transport error before connect
This fixes an issue where the transport has not been connected and a
socket error or disconnect occurs. When this happens, the endpoint has
not yet been assigned to a zap_io_thread, however, the cleanup logic
assumes that it is.

This also fixes a couple of memory leaks that were discovered during the
testing of this fix.