Add work around for thrstat issues
Presently on ldms the passive side of a connection is mishandling
the thrstat allocation and assignment of the thrstat structure
as the thread context.

This is not a fix for the real problem. The real problem is that
thread assignment to the passive endpoint is not made until accept.
This means that the thread stats are not actually kept properly.
Even though the data is collected it is stored into a thrstat
structure that is being leaked.

A second issue is that this logic is racy because the endpoint
starts out on the listen thread which is racing with the io_thread.

So this is a very long way of saying that this logic has to be
refactored, but this will be done in 4.5.2.