Fix thrstat on connection request
On CONNECT_REQUEST event, the zap endpoint does not have a thread
assigned to it yet, and hence `zap_thrstat_ctxt_set()` failed and
`thrstat` was freed. However, `thrstat` was later used after it was
freed, corrupting memory.

The `thrstat` for CONNECT_REQUEST event should actually be the `thrstat`
object of the listening endpoint, because the thread processing the
event is actually the thread of the parent endpoint. This patch get/init
thrstat accordingly.