Update ldmsd_set_deregister to receive ldmsd_plug_handle_t
When ldmsd_set_deregister() was implemented, plugins did not have access
to plugin handles. It knows only the plugin name, the code has been
involved to the point that plugins have access to plugin handles.

The patch changes the signature of ldmsd_set_deregister() to receive
ldmsd_plug_handle_t instead of the plugin name string.