Add support for msg_enable in configuration files.
The msg_enable command was previously only supported for network
clients such as ldmsd_controller.

This change also changes how messages are disabled in LDMSD. Previously,
messages were disabled in the API itself by default. This forces
clients external to LDMSD to explicitly call ldms_msg_enable() in order
to enable messages. With this change, messages are disabled in the LDMS
Daemon by default, but enabled in the API.