Fix json_stream_saampler - it should use ldmsd_stream
`json_stream_sampler` was introduced to `main` branch together with
the new `ldms_stream` (which later refactored to `ldms_msg`). However,
`json_stream_sampler` was back-ported to b4.4 which used `ldmsd_stream`.
The `json_stream_sampler` on b4.4 also added `instance_fmt` feature.

This patch makes `json_stream_sampler` on `main` to be compatible with
that on `b4.4`, and updates doc+usage.