Remove call to dlclose in plugin unload
Unmapping the shared library makes all references into the
library invalid. This includes callback functions
registered with LDMS transports, ldms_msg, and ldms_stream.
The callback functions are triggered by logic in the plugin
destructor() (e.g. ldms_xprt_close(), ldms_msg_client_close()),
to clean up resources. The plugin will recieve asynchronous CLOSE
events in the callback functions notifying the plugin that this
is the last event they can expect to receive.

A call to dlclose(), unmaps these functions resulting in a SEGFAULT
when the infrastructure attemtps to deliver the CLOSE events.
If dlclose() is to be called, it will have to be synchronized
with the completion of the callback functions. Presently many
plugins do not do this synchronization.

Synchronization logic is being designed that will make it simpler
for plugins to implement the necessary synchronization. When
this logic has been completed, and the plugins updated, the
call to dlclose() will return.