Add a maximum wait time for QUOTA credits
This option sets a timeout on the conditional wait that is
inside the transport callback handler. If the timeout expires,
the application will exit insttead of waiting for the update
indefinitely.

This option can be triggered by publishing to a peer (sitting in gdb)
and waiting for the timeout to expire.

This change also fixes up some issues discovered with the usage
message while adding the new -g option.