Fix setns() issue on slurm due to ldms_msg
`setns(fd, CLONE_NEWNS)` (mount namespace) returns `EINVAL` if the
caller is a multithreaded process. When slurm JobContainerType config
variable is set, `setns()` will be called. Since `slurm_notifier` uses
`ldms_msg` which has a thread to deliver CLIENT_CLOSE event, slurmstepd
call to `setns()` failed when `slurm_notifier` is loaded in the
plugstack.

This patch clean up the thread in `ldms_msg` at fork and respawn the
thread when it is needed.